yolo-server/requirements.txt
fastapi
uvicorn
ultralytics
pillow

